Classes and tests for bank
Added more balance sheet items and parameters to banks, firms, and
households (so we can have the model be stock-flow consistent). Further,
tweaked printing in the transactions a bit, since we have classes in the
from/to fields, and it wouldn't print nicely out-of-the-box. And tests
for banks are ready, everything seems to be working. Tests for firms and
households to follow, and then abstracting these three as much as
possible before I go into the model mechanics (runner & config &
updater).