Fixing and updater, runner
Fixes to agents and environment: checked all __del__ instances and
amended accordingly (left some that do stuff even though they're not
necessary with CPython), and added __getattr__ for clean reading of
parameters across agents and environment (though I'm wondering if I
could stop this being used as tool for writing stuff). And cleaned the
updater and runner a bit from the original black_rhino, the two will be
worked on next.