global balance checker + test, some Updater work
Added a global balance checked, the transactions of certain kind (ie
deposits) need to be the same at two agents, so I was thinking of moving
the transactions to a central list, but that creates other problems, so
I'm not sure. But if they are disjoint we need to make sure they are
always in line through updating, in addition to balance sheets of banks
being balanced. Added some updater stuff too, that'll need to be
expanded however.