Config files + fixes
Changed the config files for testing. The transactions which are read
from the config files are added to both sides now automatically, so they
do not need to be repeated in both agents' config files. When making
config files we need to keep an eye on this, i.e. put the transaction of
a given type always to one side only, e.g. the from_ side. Also checked
agents and helper for anything resembling interacting with Transactions,
should be operational now.