Abstracting and fixes
Abstracted away the __getattr__ for agents and environment. Also
abstracted the function for checking global consistency of transactions.
The code there got simplified with the generator for agents. Same
applies to the function for getting an agent through unique ID. Agents
are now wrapped in one list agents. Some typos fixed as well. Need to go
through all the tests again to see if something weird isn't happening.