Fixes to Transaction
So, first thing is that purge_accounts() on an agent level makes very
little sense. And it will be hard to control for these across agents.
Purging accounts across all agents works perfectly and shouldn't be too
much overheard, so I've deleted the individual one. Second thing is that
up until now we've been removing transactions from the books through a
deconstructor (__del__), this is not a good idea. Some weird situations
may arise when there are unknown references to the transaction, and
calling deconstructor explicitly is just bad manners. So removing
transactions from agents' books should be done through
remove_transaction() which is consistent with add_transaction() and a
better practice. I need to look through all the code to see if we have
some remains of the old practice left. Deconstructor is left as is for
safety, it doesn't hurt either way, but shouldn't be used explicitly in
the code.