Updater #1
Started with the updater, want to make it modular. Three functions
added, for interest accruals, endowing labour and removing not-selled
labour, these need to be checked and tested. Rest to follow. Note: need
to check how the delete stuff works around transactions again, just to
check if we don't have silly loops somewhere when changing arrays.