Market abstraction and tatonnement testing
Abstracted Market to BaseMarket and tested that tatonnement actually
finds a price given reasonable inputs. There should be failsafes since
otherwise it'll just run forever on a server in weird cases, to think
about. The model in Updater should be made to work with the market
clearing methods, preferably one of each, so testing is complete in this
sense.