Updater refactoring
Updater has been more or less refactored to the new double bookkeeping
version, whether it works or makes sense remains to be seen. Either way,
the bank stuff is not implemented yet, and as such interests are
breaking the balance sheets. That will be implemented next, and further
we'll take a look whether economics of it make sense.