Measurement class and some modelling
Measurement class has been added to handle writing stuff to csv files.
It works off an xml config file which specifies what should be written
and where, and accepts both static and dynamic things to be used.  It
works, some further comments, expecially about the xml config to be
provided, and all needs to be moved to an abstract class, since there
should be nothing concrete in this class. Some fiddling with the model
also happened, it now works more or less the way we described, ie
w*L=p*C, but I'll read the macro textbook, think it through and will
adjust specifics.