LOANS >> capital plus labour
Changed the macromodel to account for both labour and capital
acquisition at firm level, it was previously assumed that loans are
paying for the labour and are the capital, which works but is not very
sane for an economic model. This will require a bit of tuning of the
updater, in particular paying off loans through selling capital and so
on. This may be tricky in places, but leads to the goal that is proper
accounting of flows, maturities etc.